Convert money columns to Decimal and add DB migrations
Introduce Flask-Migrate/Alembic: baseline the existing schema and add
a money-to-numeric revision that alters wallet.balance,
transactions.total_value/price_per_share/quantity and exchange_rates.rate
from float/int to Numeric (verified value-preserving and reversible on
Postgres 16). Migrations own the schema now (create_all removed; the
admin/seed/scheduler startup is gated so `flask db` CLI is side-effect free
and bootstrap.sh runs `flask db upgrade` before gunicorn).

Serialize Decimal as JSON via a custom provider, use Decimal defaults, and
drop the as_decimal workaround so the money path is Decimal end-to-end.

+++ app/models/exchangerate.py
@@ -7,7 +7,7 @@ class ExchangeRate(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     base_currency = db.Column(db.Enum(WalletCurrencyType), nullable=False)
     target_currency = db.Column(db.Enum(WalletCurrencyType), nullable=False)
-    rate = db.Column(db.Float, nullable=False)
+    rate = db.Column(db.Numeric(18, 8), nullable=False)
     last_updated = db.Column(db.DateTime(timezone=True), default=db.func.now(), onupdate=db.func.now(), nullable=False)
